+++ src/main/java/depromeet/onepiece/feedback/command/infrastructure/FeedbackCommandRepository.java
@@ -1,4 +1,4 @@
-package depromeet.onepiece.feedback.command.domain;
+package depromeet.onepiece.feedback.command.infrastructure;
 
 import depromeet.onepiece.feedback.domain.Feedback;
 import java.util.Optional;
