--- src/main/java/depromeet/onepiece/feedback/command/presentation/response/FeedbackResponse.java
@@ -1,5 +1,0 @@
-package depromeet.onepiece.feedback.command.presentation.response;
-
-public record FeedbackResponse(
-    OverallFeedbackResponse overallFeedbackResponse,
-    ProjectFeedbackResponse projectFeedbackResponse) {}
