+++ src/main/java/depromeet/onepiece/feedback/domain/FeedbackStatus.java
@@ -6,6 +6,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum FeedbackStatus {
+  PENDING("PENDING"),
   COMPLETE("COMPLETE"),
   IN_PROGRESS("IN_PROGRESS"),
   ERROR("ERROR");
