+++ src/main/java/depromeet/onepiece/feedback/query/infrastructure/FeedbackQueryRepositoryImpl.java
@@ -2,7 +2,6 @@ package depromeet.onepiece.feedback.query.infrastructure;
 
 import depromeet.onepiece.feedback.domain.Feedback;
 import depromeet.onepiece.feedback.query.application.exception.FeedbackNotFoundException;
-import depromeet.onepiece.feedback.query.domain.FeedbackQueryRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
